+++ modules/rd/controllers/ContainerController.php
@@ -37,6 +37,9 @@ class ContainerController extends Controller
      */
     public function actionIndex()
     {
+        if(!Yii::$app->user->can("warehouse_list"))
+            throw new ForbiddenHttpException('Usted no tiene permiso ver esta vista');
+
         $searchModel = new ContainerSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -54,6 +57,9 @@ class ContainerController extends Controller
      */
     public function actionView($id)
     {
+        if(!Yii::$app->user->can("warehouse_view"))
+            throw new ForbiddenHttpException('Usted no tiene permiso ver esta vista');
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -66,6 +72,9 @@ class ContainerController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can("warehouse_create"))
+            throw new ForbiddenHttpException('Usted no tiene permiso ver esta vista');
+
         $model = new Container();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -86,6 +95,9 @@ class ContainerController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can("warehouse_update"))
+            throw new ForbiddenHttpException('Usted no tiene permiso ver esta vista');
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -106,6 +118,9 @@ class ContainerController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can("warehouse_delete"))
+            throw new ForbiddenHttpException('Usted no tiene permiso ver esta vista');
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
